app/requirements.txt
fastapi==0.68.2
mangum==0.12.2
uvicorn[standard]==0.15.0
boto3
pycryptodome==3.14.1